big refactorings
refactored the transactor to no longer wait for the TX confirmations as
this is now tracked in the profit tracker.
added gas price querier package  and interface.
added more metrics to the profit tracker - tracks failed and success TX.
refactored to use ctx instead of SubmitCanceler
refactored to use differenc submiters for the different contracts as the
signatures are different
renamed package to use underscore instead of camelCase to avoid name
collision with variables
simplified the default gas estimator with hard coded values as it was
too complicated with the eth client refund logic
moved the submit metrics in the reward tracker - failed, success, submit
values.
increased the web fetch timeout to 10 secs

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>