Open sourcing it again
added Get Data command to show the last submitted values
added dispute list command

added congecko for ZRXUSDT

Added a command to encrypt the file on disk and when ask for a password
when trying to load an encrypted file

Added unlock fee command

allow setting  custom gas limits for all commands

upgraded all submits to use eip-1559

Converted the k8s deployment to use helm charts thanks to
@ImagineBeingAtComputers

removee inactive DATA ids and removed all apis fron the index file that
are not needed.

Renamed miners into reporters in all places

upgrade geth to 1.10.7

added tally list command

simplified the re-org event handling

removed mesosphere as it is not used right now

in the index file changed the format to use param for value and param
for time to be able to check for very old timestamps

Increased the default index intervals to 1 minute as most API don't
return new data that frequently.

Exported the Reorg event wait to be global constant so that all
components wait times can be updated in a single place

make prometheus deployment optional

Now that all metrics are saved directly in the Db don't need prometheus

Added decrypt command to allow restoring encrypted files

refactored the default transactor for eip1559

Added global validation for all commands that write to the blockchain.
This ensures that will not send a transaction by mistace with very high
TX fees.

renamed BigIntToFloat to BigIntToFloatDiv to indicate that it performs
division and FloatToBigInt to FloatToBigIntMul respectively.

Renamed GasTipFee cli arg to GasTip

Add prompt for values when doing manual submit

Make the net ID call only once when the client is created

this reduces the network calls a bit
increases the verbosity a bit as the netID needs to be passed around,
but it is not too bad and a good tradeoff

rename helm reference to use command instead of mode or cmd

no longer generate bindings for balancer and uniswap

These are not used for now.

Added test contract bindings generate target in the Makefile

This is used in some custom scripts to mint new tokens

Fix the math percentage diff formula and test

don't create manualData config map if file don't exist

fixeda panic bug in the math lib when the devider is 1

improve the submit command

auto add the submit values from db if available or otherwise asks for
manual values

add proxy flag to the manual submit command

usefull for testing TX that are submitted to proxy contracts

added generic events tracker package to reduce code duplication

Also added cli command to query and subscribe to logs

hex format validation for all commands that take address

refactor flags to be more explicit for Account and Contract

In helm - make the loadbalancer disabled by default

install the service as node port by default as the node balancer is paid
and not really needed

Also in grafana don't install default prometheus datasource when
prometheus is not enabled

add another arg to make the account arg optional

Make all event names as constants

Extend the contract caller interface with Abi and UnpackLog

added a flag in the submit cmd to allow skipping the confirm

this is useful for testing

execute url env expansion before every request

Before this time variables were expanded only at startup

Add contract params options

When initiating a contract instance can now specify variables like
dispute voting window and unlock window. this is useful for testing when
want to have quick iteration when testing the execution scheduling

enable strict parsing for the index file

fix the url env expander logic in the bravenew coin api
Didn't setup the  api keys properly

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>