don't use os get and set env
this is to make it more secure when running in a shared env
for example when using an encrypted file it will load only in memory and
not in the env vars. This way it is more secure to give access to
someone to the machine without exposing the private key.

Signed-off-by: Krasi Georgiev <8903888+krasi-georgiev@users.noreply.github.com>